+++ app/src/main/java/com/codesoom/assignment/services/product/exception/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment;
+package com.codesoom.assignment.services.product.exception;
 
 public class ProductNotFoundException extends RuntimeException {
     public ProductNotFoundException(Long id) {
