+++ app/src/test/java/com/codesoom/assignment/controllers/ProductControllerTest.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment.controllers;
 
-import com.codesoom.assignment.ProductNotFoundException;
+import com.codesoom.assignment.services.product.exception.ProductNotFoundException;
 import com.codesoom.assignment.services.product.ProductService;
 import com.codesoom.assignment.controllers.product.ProductController;
 import com.codesoom.assignment.services.product.domain.Product;
